refactor(types): Moved `collection` types to separate file.
