+++ frontend/src/components/Hero/Hero.astro
@@ -1,5 +1,6 @@
 ---
-import type { CTA, SocialIcon } from "@/types/common";
+import type { CTA } from "@/types/common";
+import type { SocialIcon } from "@/types/collections";
 
 import { CTAs, Heading } from "@/components/Hero";
 import SocialButton from "@/components/SocialButton.astro";
