--- frontend/src/components/Link.astro
@@ -1,9 +1,0 @@
----
-import type { Link } from "@/types/common.ts";
-
-const { href, styles } = Astro.props as Link;
----
-
-<a href={href} class={styles}>
-  <slot />
-</a>
