+++ frontend/src/components/SocialButton.astro
@@ -1,5 +1,5 @@
 ---
-import type { SocialIcon } from "@/types/common";
+import type { SocialIcon } from "@/types/collections";
 
 type Props = SocialIcon & {
   invert?: boolean;
