+++ frontend/src/layouts/BaseLayout.astro
@@ -4,7 +4,7 @@ import PageBg from "@/assets/imgs/page-bg.jpg";
 import { Content as AboutContent } from "@/content/about.md";
 import { Head, Scripts } from "@/components/global";
 import Sidebar from "@/components/Sidebar";
-import type { SocialIcon } from "@/types/common";
+import type { SocialIcon } from "@/types/collections";
 
 type Props = {
   title: string;
