+++ frontend/src/pages/index.astro
@@ -4,7 +4,8 @@ import { Navbar } from "@/components/Navbar";
 
 import Layout from "@/layouts/Homepage.astro";
 
-import { NavbarLinks, HeroCTAs, SocialIcons } from "@/data/homepage";
+import { NavbarLinks, HeroCTAs } from "@/data/homepage";
+import type { SocialIcon } from "@/types/collections";
 
 // const req = await fetch("");
 // const data = await req.json();
