+++ frontend/src/pages/journal.astro
@@ -3,7 +3,7 @@ import Layout from "@/layouts/BaseLayout.astro";
 
 import { SocialIcons } from "@/data/homepage";
 import { ArticleSection } from "@/components/Portfolio";
-
+import type { Project, SocialIcon } from "@/types/collections";
 import type { Project } from "@/types/common";
 
 import { getCollection } from "astro:content";
