+++ frontend/src/pages/journal/[projectSlug]/[articleSlug].astro
@@ -3,7 +3,7 @@ import Layout from "@/layouts/ProjectSingle.astro";
 import { SingleHeader } from "@/components/Portfolio";
 
 import { NavbarLinks, SocialIcons } from "@/data/homepage";
-import type { Article } from "@/types/common";
+import type { Article } from "@/types/collections";
 import { formatArticleCollection } from "@/utils";
 
 import { getCollection } from "astro:content";
